chore: delete comments and files chore: rename files and content to be more consistent
